Fix `lsp_hover_max_lines` to deal with diagnostics
* Made `lsp_hover_max_lines` control the size of the entire hover box
and not just the info portion. When the output for the hover box
exceeds `lsp_hover_max_lines` the info and diagnostic lines are
truncated to so that the hover box does not exceed
`lsp_hover_max_lines`. All the info lines will be truncated before any
diagnostic lines are truncated.
* Added `lsp_hover_max_info_lines` to control the max info lines in the
hover box.
* Added `lsp_hover_max_diagnostic_lines` to control the max diagnostic
lines in the hover box.